update nav component using css bootstrap3
I learned a little bigt more about how to use css bootstrap and applied that knowledge to make a fast and easy to create simple nav component which floats to the left of my resume web site making it look nice. Thus my quest to improve the layout of this page continues as i seek to master the various design flow options available in the angular library but i know that i should alsdo learn about css bootstrap and generic templates to put in to the css files of my components.